WIP: move some things to url-utils.
