+++ report/tsconfig.json
@@ -14,6 +14,8 @@
   "include": [
     "**/*.js",
     "types/**/*.d.ts",
+    "../core/lib/url-utils.js",
+    "../core/lib/lh-error.js",
   ],
   "exclude": [
     "generator/**/*.js",
